+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -5,6 +5,7 @@ import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.UserService;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.UUID;
 
 
@@ -38,19 +39,25 @@ public class BasicUserService implements UserService {
 
     @Override
     public void delete(UUID userId) {
+        if(userRepository.findId(userId)==null)
+            throw new NoSuchElementException(userId + "를 찾을 수 없습니다.");
         userRepository.remove(userId);
     }
 
     @Override
     public User update(UUID userId, String name, String nickname, String email) {
         User user = findId(userId);
+        if(user==null)
+            throw new NoSuchElementException(userId + "를 찾을 수 없습니다.");
         user.update(name, nickname, email);
         return user;
     }
 
     @Override
-    public User findId(UUID userID) {
-        return userRepository.findId(userID);
+    public User findId(UUID userId) {
+        if(userRepository.findId(userId)==null)
+            throw new NoSuchElementException(userId + "를 찾을 수 없습니다.");
+        return userRepository.findId(userId);
     }
 
 }
